+++ src/router/PrivateRoute.tsx
@@ -1,0 +1,8 @@
+import { Navigate, Outlet } from 'react-router-dom';
+
+function PrivateRoute() {
+	const isAuthenticated = localStorage.getItem('accessToken');
+	return isAuthenticated ? <Outlet /> : <Navigate to="/" />;
+}
+
+export default PrivateRoute;
